Working edit

+++ pkgs/fajli/src/fajli/commands.py
@@ -37,7 +37,7 @@ def get(
         print(f"File {file} is not a part of fajli")
         sys.exit(1)
 
-def set(
+def update(
     fajli: Fajli,
     out_file: Path,
     in_file = os.sys.stdin,
@@ -58,6 +58,35 @@ def set(
                         print("Enter file contents (ctrl-d when done):")
                     with open(workdir / file_cfg['path'], 'w') as f:
                         f.writelines(in_file)
+                    
+                    if is_encrypted:
+                        (workdir / f'{file_cfg['path']}.age').unlink()
+                        fajli.encrypt(workdir, folder_cfg)
+
+                    file_found = True
+                    break
+            
+            if file_found:
+                break
+
+def edit(
+    fajli: Fajli,
+    out_file: Path,
+) -> int:
+    with fajli.transactional() as workdir:
+        file_found = False
+
+        for folder, folder_cfg in fajli:
+            fajli.decrypt(workdir, folder_cfg)
+
+            for file_name, file_cfg in folder_cfg['files'].items():
+                is_encrypted = file_cfg['age']['enable']
+                matches_age = str(out_file.absolute()).endswith(f'{file_cfg['path']}.age')
+                matches = str(out_file.absolute()).endswith(file_cfg['path'])
+
+                if is_encrypted and matches_age or matches:
+                    editor = os.environ.get('EDITOR') or sys.exit("Missing EDITOR environment")
+                    subprocess.call(args=[ editor, workdir / file_cfg['path']])
                     
                     if is_encrypted:
                         (workdir / f'{file_cfg['path']}.age').unlink()
